Bugfix: Store Thycotic object in an instance scope variable, not global.
Storing the 'Thycotic' object in a GLOBAL scoped variable is dangerous
in threaded code. Storing it as an INSTANCE variable is safer, and
provides the same functionality of making sure we create a single
Thycotic object and use it for as long as possible before having to
create another one.

Change-Id: I04c39889ee3176f8c1734cc3277bae78b6026821